move `HashDiscarded` to `exceptions` module
